@@ -6,14 +6,14 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
   cupertino_icons: ^1.0.0
-  modal_progress_hud: ^0.1.3
+  modal_progress_hud_nsn: ^0.1.0-nullsafety-1
   numberpicker: ^2.1.1
   image_picker: ^0.8.0+4
   firebase_storage: ^10.2.9
