@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:tinder_app_flutter/data/db/entity/chat.dart';
@@ -18,15 +20,14 @@ class UserProvider extends ChangeNotifier {
   FirebaseDatabaseSource _databaseSource = FirebaseDatabaseSource();
 
   bool isLoading = false;
-  AppUser _user;
+  AppUser? _user;
 
-  Future<AppUser> get user => _getUser();
+  Future<AppUser?> get user => _getUser();
 
-  Future<Response> loginUser(String email, String password,
-      GlobalKey<ScaffoldState> errorScaffoldKey) async {
+  Future<Response> loginUser(String email, String password, GlobalKey<ScaffoldState> errorScaffoldKey) async {
     Response<dynamic> response = await _authSource.signIn(email, password);
     if (response is Success<UserCredential>) {
-      String id = response.value.user.uid;
+      String id = response.value.user!.uid;
       SharedPreferencesUtil.setUserId(id);
     } else if (response is Error) {
       showSnackBar(errorScaffoldKey, response.message);
@@ -34,22 +35,17 @@ class UserProvider extends ChangeNotifier {
     return response;
   }
 
-  Future<Response> registerUser(UserRegistration userRegistration,
-      GlobalKey<ScaffoldState> errorScaffoldKey) async {
-    Response<dynamic> response = await _authSource.register(
-        userRegistration.email, userRegistration.password);
+  Future<Response> registerUser(UserRegistration userRegistration, GlobalKey<ScaffoldState> errorScaffoldKey) async {
+    Response<dynamic> response = await _authSource.register(userRegistration.email, userRegistration.password);
     if (response is Success<UserCredential>) {
-      String id = (response as Success<UserCredential>).value.user.uid;
-      response = await _storageSource.uploadUserProfilePhoto(
-          userRegistration.localProfilePhotoPath, id);
+      String id = response.value.user!.uid;
+      response = await (_storageSource.uploadUserProfilePhoto(userRegistration.localProfilePhotoPath, id)
+          as FutureOr<Success<UserCredential>>);
 
       if (response is Success<String>) {
-        String profilePhotoUrl = response.value;
-        AppUser user = AppUser(
-            id: id,
-            name: userRegistration.name,
-            age: userRegistration.age,
-            profilePhotoPath: profilePhotoUrl);
+        String profilePhotoUrl = response.value as String;
+        AppUser user =
+            AppUser(id: id, name: userRegistration.name, age: userRegistration.age, profilePhotoPath: profilePhotoUrl);
         _databaseSource.addUser(user);
         SharedPreferencesUtil.setUserId(id);
         _user = _user;
@@ -60,23 +56,21 @@ class UserProvider extends ChangeNotifier {
     return response;
   }
 
-  Future<AppUser> _getUser() async {
+  Future<AppUser?> _getUser() async {
     if (_user != null) return _user;
-    String id = await SharedPreferencesUtil.getUserId();
+    String? id = await SharedPreferencesUtil.getUserId();
     _user = AppUser.fromSnapshot(await _databaseSource.getUser(id));
     return _user;
   }
 
-  void updateUserProfilePhoto(
-      String localFilePath, GlobalKey<ScaffoldState> errorScaffoldKey) async {
+  void updateUserProfilePhoto(String localFilePath, GlobalKey<ScaffoldState> errorScaffoldKey) async {
     isLoading = true;
     notifyListeners();
-    Response<dynamic> response =
-        await _storageSource.uploadUserProfilePhoto(localFilePath, _user.id);
+    Response<dynamic> response = await _storageSource.uploadUserProfilePhoto(localFilePath, _user!.id);
     isLoading = false;
     if (response is Success<String>) {
-      _user.profilePhotoPath = response.value;
-      _databaseSource.updateUser(_user);
+      _user!.profilePhotoPath = response.value;
+      _databaseSource.updateUser(_user!);
     } else if (response is Error) {
       showSnackBar(errorScaffoldKey, response.message);
     }
@@ -84,8 +78,8 @@ class UserProvider extends ChangeNotifier {
   }
 
   void updateUserBio(String newBio) {
-    _user.bio = newBio;
-    _databaseSource.updateUser(_user);
+    _user!.bio = newBio;
+    _databaseSource.updateUser(_user!);
     notifyListeners();
   }
 
@@ -94,16 +88,15 @@ class UserProvider extends ChangeNotifier {
     await SharedPreferencesUtil.removeUserId();
   }
 
-  Future<List<ChatWithUser>> getChatsWithUser(String userId) async {
+  Future<List<ChatWithUser>> getChatsWithUser(String? userId) async {
     var matches = await _databaseSource.getMatches(userId);
     List<ChatWithUser> chatWithUserList = [];
 
     for (var i = 0; i < matches.size; i++) {
       Match match = Match.fromSnapshot(matches.docs[i]);
-      AppUser matchedUser =
-          AppUser.fromSnapshot(await _databaseSource.getUser(match.id));
+      AppUser matchedUser = AppUser.fromSnapshot(await _databaseSource.getUser(match.id));
 
-      String chatId = compareAndCombineIds(match.id, userId);
+      String chatId = compareAndCombineIds(match.id!, userId!);
 
       Chat chat = Chat.fromSnapshot(await _databaseSource.getChat(chatId));
       ChatWithUser chatWithUser = ChatWithUser(chat, matchedUser);
