@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:tinder_app_flutter/data/db/entity/app_user.dart';
@@ -11,23 +13,23 @@ import 'package:tinder_app_flutter/util/utils.dart';
 class MatchedScreen extends StatelessWidget {
   static const String id = 'matched_screen';
 
-  final String myProfilePhotoPath;
-  final String myUserId;
-  final String otherUserProfilePhotoPath;
-  final String otherUserId;
+  final String? myProfilePhotoPath;
+  final String? myUserId;
+  final String? otherUserProfilePhotoPath;
+  final String? otherUserId;
 
   MatchedScreen(
-      {@required this.myProfilePhotoPath,
-      @required this.myUserId,
-      @required this.otherUserProfilePhotoPath,
-      @required this.otherUserId});
+      {required this.myProfilePhotoPath,
+      required this.myUserId,
+      required this.otherUserProfilePhotoPath,
+      required this.otherUserId});
 
   void sendMessagePressed(BuildContext context) async {
-    AppUser user = await Provider.of<UserProvider>(context, listen: false).user;
+    AppUser user = await (Provider.of<UserProvider>(context, listen: false).user as FutureOr<AppUser>);
 
     Navigator.pop(context);
     Navigator.pushNamed(context, ChatScreen.id, arguments: {
-      "chat_id": compareAndCombineIds(myUserId, otherUserId),
+      "chat_id": compareAndCombineIds(myUserId!, otherUserId!),
       "user_id": user.id,
       "other_user_id": otherUserId
     });
@@ -54,10 +56,7 @@ class MatchedScreen extends StatelessWidget {
               Container(
                 child: Row(
                   mainAxisAlignment: MainAxisAlignment.spaceAround,
-                  children: [
-                    Portrait(imageUrl: myProfilePhotoPath),
-                    Portrait(imageUrl: otherUserProfilePhotoPath)
-                  ],
+                  children: [Portrait(imageUrl: myProfilePhotoPath), Portrait(imageUrl: otherUserProfilePhotoPath)],
                 ),
               ),
               Column(
