@@ -6,7 +6,7 @@ import 'package:tinder_app_flutter/ui/widgets/image_portrait.dart';
 class AddPhotoScreen extends StatefulWidget {
   final Function(String) onPhotoChanged;
 
-  AddPhotoScreen({@required this.onPhotoChanged});
+  AddPhotoScreen({required this.onPhotoChanged});
 
   @override
   _AddPhotoScreenState createState() => _AddPhotoScreenState();
@@ -14,7 +14,7 @@ class AddPhotoScreen extends StatefulWidget {
 
 class _AddPhotoScreenState extends State<AddPhotoScreen> {
   final picker = ImagePicker();
-  String _imagePath;
+  String? _imagePath;
 
   Future pickImageFromGallery() async {
     final pickedFile = await picker.getImage(source: ImageSource.gallery);
