@@ -4,7 +4,7 @@ import 'package:tinder_app_flutter/ui/widgets/bordered_text_field.dart';
 class NameScreen extends StatelessWidget {
   final Function(String) onChanged;
 
-  NameScreen({@required this.onChanged});
+  NameScreen({required this.onChanged});
 
   @override
   Widget build(BuildContext context) {
