@@ -22,11 +22,11 @@ class MatchScreen extends StatefulWidget {
 class _MatchScreenState extends State<MatchScreen> {
   final FirebaseDatabaseSource _databaseSource = FirebaseDatabaseSource();
   final _scaffoldKey = GlobalKey<ScaffoldState>();
-  List<String> _ignoreSwipeIds;
+  List<String?>? _ignoreSwipeIds;
 
-  Future<AppUser> loadPerson(String myUserId) async {
+  Future<AppUser?> loadPerson(String? myUserId) async {
     if (_ignoreSwipeIds == null) {
-      _ignoreSwipeIds = List<String>();
+      List<String?> _ignoreSwipeIds = [];
       var swipes = await _databaseSource.getSwipes(myUserId);
       for (var i = 0; i < swipes.size; i++) {
         Swipe swipe = Swipe.fromSnapshot(swipes.docs[i]);
@@ -45,13 +45,13 @@ class _MatchScreenState extends State<MatchScreen> {
 
   void personSwiped(AppUser myUser, AppUser otherUser, bool isLiked) async {
     _databaseSource.addSwipedUser(myUser.id, Swipe(otherUser.id, isLiked));
-    _ignoreSwipeIds.add(otherUser.id);
+    _ignoreSwipeIds!.add(otherUser.id);
 
     if (isLiked == true) {
       if (await isMatch(myUser, otherUser) == true) {
         _databaseSource.addMatch(myUser.id, Match(otherUser.id));
         _databaseSource.addMatch(otherUser.id, Match(myUser.id));
-        String chatId = compareAndCombineIds(myUser.id, otherUser.id);
+        String chatId = compareAndCombineIds(myUser.id!, otherUser.id!);
         _databaseSource.addChat(Chat(chatId, null));
 
         Navigator.pushNamed(context, MatchedScreen.id, arguments: {
@@ -66,8 +66,7 @@ class _MatchScreenState extends State<MatchScreen> {
   }
 
   Future<bool> isMatch(AppUser myUser, AppUser otherUser) async {
-    DocumentSnapshot swipeSnapshot =
-        await _databaseSource.getSwipe(otherUser.id, myUser.id);
+    DocumentSnapshot swipeSnapshot = await _databaseSource.getSwipe(otherUser.id, myUser.id);
     if (swipeSnapshot.exists) {
       Swipe swipe = Swipe.fromSnapshot(swipeSnapshot);
 
@@ -84,25 +83,18 @@ class _MatchScreenState extends State<MatchScreen> {
         key: _scaffoldKey,
         body: Container(child: Consumer<UserProvider>(
           builder: (context, userProvider, child) {
-            return FutureBuilder<AppUser>(
+            return FutureBuilder<AppUser?>(
               future: userProvider.user,
               builder: (context, userSnapshot) {
                 return CustomModalProgressHUD(
-                  inAsyncCall:
-                      userProvider.user == null || userProvider.isLoading,
+                  inAsyncCall: userProvider.user == null || userProvider.isLoading,
                   child: (userSnapshot.hasData)
-                      ? FutureBuilder<AppUser>(
-                          future: loadPerson(userSnapshot.data.id),
+                      ? FutureBuilder<AppUser?>(
+                          future: loadPerson(userSnapshot.data!.id),
                           builder: (context, snapshot) {
-                            if (snapshot.connectionState ==
-                                    ConnectionState.done &&
-                                !snapshot.hasData) {
+                            if (snapshot.connectionState == ConnectionState.done && !snapshot.hasData) {
                               return Center(
-                                child: Container(
-                                    child: Text('No users',
-                                        style: Theme.of(context)
-                                            .textTheme
-                                            .headline4)),
+                                child: Container(child: Text('No users', style: Theme.of(context).textTheme.headline4)),
                               );
                             }
                             if (!snapshot.hasData) {
@@ -116,39 +108,28 @@ class _MatchScreenState extends State<MatchScreen> {
                                 padding: EdgeInsets.all(12.0),
                                 child: Container(
                                   child: Column(
-                                    crossAxisAlignment:
-                                        CrossAxisAlignment.center,
+                                    crossAxisAlignment: CrossAxisAlignment.center,
                                     children: [
                                       SwipeCard(person: snapshot.data),
                                       Expanded(
                                         child: Container(
-                                          margin: EdgeInsets.symmetric(
-                                              horizontal: 45),
+                                          margin: EdgeInsets.symmetric(horizontal: 45),
                                           child: Align(
                                             alignment: Alignment.center,
                                             child: Row(
-                                              mainAxisAlignment:
-                                                  MainAxisAlignment
-                                                      .spaceBetween,
+                                              mainAxisAlignment: MainAxisAlignment.spaceBetween,
                                               children: [
                                                 RoundedIconButton(
                                                   onPressed: () {
-                                                    personSwiped(
-                                                        userSnapshot.data,
-                                                        snapshot.data,
-                                                        false);
+                                                    personSwiped(userSnapshot.data!, snapshot.data!, false);
                                                   },
                                                   iconData: Icons.clear,
-                                                  buttonColor:
-                                                      kColorPrimaryVariant,
+                                                  buttonColor: kColorPrimaryVariant,
                                                   iconSize: 30,
                                                 ),
                                                 RoundedIconButton(
                                                   onPressed: () {
-                                                    personSwiped(
-                                                        userSnapshot.data,
-                                                        snapshot.data,
-                                                        true);
+                                                    personSwiped(userSnapshot.data!, snapshot.data!, true);
                                                   },
                                                   iconData: Icons.favorite,
                                                   iconSize: 30,
