@@ -7,18 +7,18 @@ class ChatListTile extends StatelessWidget {
   final ChatWithUser chatWithUser;
   final Function onTap;
   final Function onLongPress;
-  final String myUserId;
+  final String? myUserId;
 
   ChatListTile(
-      {@required this.chatWithUser,
-      @required this.onTap,
-      @required this.onLongPress,
-      @required this.myUserId});
+      {required this.chatWithUser,
+      required this.onTap,
+      required this.onLongPress,
+      required this.myUserId});
 
   @override
   Widget build(BuildContext context) {
     return GestureDetector(
-      onTap: onTap,
+      onTap: onTap as void Function()?,
       onLongPress: () {},
       child: Container(
         height: 60,
@@ -29,7 +29,7 @@ class ChatListTile extends StatelessWidget {
               child: CircleAvatar(
                 radius: 50,
                 backgroundImage:
-                    NetworkImage(chatWithUser.user.profilePhotoPath),
+                    NetworkImage(chatWithUser.user.profilePhotoPath!),
               ),
               decoration: BoxDecoration(
                 shape: BoxShape.circle,
@@ -51,11 +51,11 @@ class ChatListTile extends StatelessWidget {
   }
 
   bool isLastMessageMyText() {
-    return chatWithUser.chat.lastMessage.senderId == myUserId;
+    return chatWithUser.chat.lastMessage!.senderId == myUserId;
   }
 
   bool isLastMessageSeen() {
-    if (chatWithUser.chat.lastMessage.seen == false &&
+    if (chatWithUser.chat.lastMessage!.seen == false &&
         isLastMessageMyText() == false) {
       return false;
     }
@@ -68,7 +68,7 @@ class ChatListTile extends StatelessWidget {
       children: [
         Expanded(
           child: Text(
-            chatWithUser.user.name,
+            chatWithUser.user.name!,
             maxLines: 1,
             overflow: TextOverflow.ellipsis,
             style: TextStyle(fontSize: 16),
@@ -79,7 +79,7 @@ class ChatListTile extends StatelessWidget {
                 chatWithUser.chat.lastMessage == null
                     ? ''
                     : convertEpochMsToDateTime(
-                        chatWithUser.chat.lastMessage.epochTimeMs),
+                        chatWithUser.chat.lastMessage!.epochTimeMs!),
                 textAlign: TextAlign.center,
                 style: TextStyle(fontSize: 12))),
       ],
@@ -97,7 +97,7 @@ class ChatListTile extends StatelessWidget {
               chatWithUser.chat.lastMessage == null
                   ? "Write something!"
                   : ((isLastMessageMyText() ? "You: " : "") +
-                      chatWithUser.chat.lastMessage.text),
+                      chatWithUser.chat.lastMessage!.text!),
               maxLines: 1,
               overflow: TextOverflow.ellipsis,
               style: TextStyle(fontSize: 14),
