@@ -3,16 +3,16 @@ import 'package:tinder_app_flutter/util/constants.dart';
 import 'package:tinder_app_flutter/util/utils.dart';
 
 class MessageBubble extends StatelessWidget {
-  final int epochTimeMs;
-  final String text;
+  final int? epochTimeMs;
+  final String? text;
   final bool isSenderMyUser;
-  final bool includeTime;
+  final bool? includeTime;
 
   MessageBubble(
-      {@required this.epochTimeMs,
-      @required this.text,
-      @required this.isSenderMyUser,
-      @required this.includeTime});
+      {required this.epochTimeMs,
+      required this.text,
+      required this.isSenderMyUser,
+      required this.includeTime});
 
   @override
   Widget build(BuildContext context) {
@@ -22,13 +22,13 @@ class MessageBubble extends StatelessWidget {
         crossAxisAlignment:
             isSenderMyUser ? CrossAxisAlignment.end : CrossAxisAlignment.start,
         children: <Widget>[
-          this.includeTime
+          this.includeTime!
               ? Opacity(
                   opacity: 0.4,
                   child: Container(
-                    child: Text(convertEpochMsToDateTime(epochTimeMs),
+                    child: Text(convertEpochMsToDateTime(epochTimeMs!),
                         textAlign: TextAlign.center,
-                        style: Theme.of(context).textTheme.bodyText1.copyWith(
+                        style: Theme.of(context).textTheme.bodyText1!.copyWith(
                             fontSize: 14, fontWeight: FontWeight.normal)),
                     width: double.infinity,
                   ),
@@ -45,8 +45,8 @@ class MessageBubble extends StatelessWidget {
               child: Padding(
                 padding: EdgeInsets.symmetric(vertical: 10.0, horizontal: 20.0),
                 child: Text(
-                  text,
-                  style: Theme.of(context).textTheme.bodyText1.copyWith(
+                  text!,
+                  style: Theme.of(context).textTheme.bodyText1!.copyWith(
                       color: isSenderMyUser ? kSecondaryColor : Colors.black,
                       fontWeight: FontWeight.normal),
                 ),
