@@ -3,9 +3,9 @@ import 'package:tinder_app_flutter/util/constants.dart';
 
 class RoundedButton extends StatelessWidget {
   final String text;
-  final Function onPressed;
+  final Function? onPressed;
 
-  RoundedButton({@required this.text, @required this.onPressed});
+  RoundedButton({required this.text, required this.onPressed});
 
   @override
   Widget build(BuildContext context) {
@@ -18,7 +18,7 @@ class RoundedButton extends StatelessWidget {
         elevation: 0,
         shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(25)),
         child: Text(text, style: Theme.of(context).textTheme.button),
-        onPressed: onPressed,
+        onPressed: onPressed as void Function()?,
       ),
     );
   }
