@@ -5,11 +5,11 @@ class RoundedIconButton extends StatelessWidget {
   final IconData iconData;
   final double iconSize;
   final paddingReduce;
-  final Color buttonColor;
+  final Color? buttonColor;
 
   RoundedIconButton(
-      {@required this.onPressed,
-      @required this.iconData,
+      {required this.onPressed,
+      required this.iconData,
       this.iconSize = 30,
       this.buttonColor,
       this.paddingReduce = 0});
@@ -21,7 +21,7 @@ class RoundedIconButton extends StatelessWidget {
       minWidth: 0,
       elevation: 5,
       color: buttonColor != null ? buttonColor : Theme.of(context).buttonColor,
-      onPressed: onPressed,
+      onPressed: onPressed as void Function()?,
       padding: EdgeInsets.all((iconSize / 2) - paddingReduce),
       child: Icon(iconData, size: iconSize),
       shape: CircleBorder(),
